Stub out new custom node tests.
