+++ src/VisualizationTests/HelixWatch3DViewModelTests.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Windows;
-using System.Windows.Media;
 using System.Windows.Media.Media3D;
 using System.Xml;
 using SystemTestServices;
@@ -12,7 +11,6 @@ using Dynamo;
 using Dynamo.Controls;
 using Dynamo.Models;
 using Dynamo.Nodes;
-using Dynamo.Selection;
 using Dynamo.Tests;
 using Dynamo.UI;
 using Dynamo.Utilities;
@@ -20,7 +18,6 @@ using Dynamo.ViewModels;
 using Dynamo.Wpf.ViewModels.Watch3D;
 using DynamoCoreWpfTests.Utility;
 using DynamoShapeManager;
-using HelixToolkit.Wpf;
 using HelixToolkit.Wpf.SharpDX;
 using NUnit.Framework;
 using SharpDX;
@@ -95,12 +92,12 @@ namespace WpfVisualizationTests
     [TestFixture]
     public class HelixWatch3DViewModelTests : VisualizationTest
     {
-        private IEnumerable<Model3D> BackgroundPreviewGeometry
+        protected IEnumerable<Model3D> BackgroundPreviewGeometry
         {
             get { return ((HelixWatch3DViewModel)ViewModel.BackgroundPreviewViewModel).SceneItems; }
-        } 
+        }
 
-        private Watch3DView BackgroundPreview
+        protected Watch3DView BackgroundPreview
         {
             get
             {
@@ -433,76 +430,6 @@ namespace WpfVisualizationTests
 
         #endregion
 
-        #region custom node tests
-
-        [Test]
-        public void CustomNodes_Render()
-        {
-            CustomNodeInfo info;
-            Assert.IsTrue(
-                ViewModel.Model.CustomNodeManager.AddUninitializedCustomNode(
-                    Path.Combine(
-                        GetTestDirectory(ExecutingDirectory),
-                        @"core\visualization\Points.dyf"),
-                    true,
-                    out info));
-
-            OpenVisualizationTest("ASM_customNode.dyn");
-
-            var ws = ViewModel.Model.CurrentWorkspace as HomeWorkspaceModel;
-
-            //ensure that we have some visualizations
-            Assert.Greater(BackgroundPreviewGeometry.TotalPoints(), 0);
-        }
-
-        [Test]
-        public void CustomNode_DoesNotRender()
-        {
-            // Regression test for defect http://adsk-oss.myjetbrains.com/youtrack/issue/MAGN-5165
-            // To verify when some geometry nodes are converted to custom node,
-            // their render packages shouldn't be carried over to custom work
-            // space.
-            var model = ViewModel.Model;
-
-            OpenVisualizationTest("visualize_line_incustom.dyn");
-
-            var ws = ViewModel.Model.CurrentWorkspace as HomeWorkspaceModel;
-
-            Assert.AreEqual(1, BackgroundPreviewGeometry.TotalCurves());
-
-            // Convert a DSFunction node Line.ByPointDirectionLength to custom node.
-            var workspace = model.CurrentWorkspace;
-            var node = workspace.Nodes.OfType<DSFunction>().First();
-
-            List<NodeModel> selectionSet = new List<NodeModel>() { node };
-            var customWorkspace = model.CustomNodeManager.Collapse(
-                selectionSet.AsEnumerable(),
-                model.CurrentWorkspace,
-                true,
-                new FunctionNamePromptEventArgs
-                {
-                    Category = "Testing",
-                    Description = "",
-                    Name = "__VisualizationTest__",
-                    Success = true
-                }) as CustomNodeWorkspaceModel;
-            ViewModel.HomeSpace.Run();
-
-            // Switch to custom workspace
-            model.OpenCustomNodeWorkspace(customWorkspace.CustomNodeId);
-            var customSpace = ViewModel.Model.CurrentWorkspace;
-
-            // Select that node
-            DynamoSelection.Instance.Selection.Add(node);
-
-            // No preview in the background
-            Assert.AreEqual(0, BackgroundPreviewGeometry.TotalPoints());
-            Assert.AreEqual(0, BackgroundPreviewGeometry.TotalCurves());
-            Assert.AreEqual(0, BackgroundPreviewGeometry.TotalMeshes());
-        }
-
-        #endregion
-
         #region watch 3d tests
 
         [Test]
@@ -705,7 +632,7 @@ namespace WpfVisualizationTests
             Assert.True(BackgroundPreviewGeometry.HasAnyColorMappedMeshes());
         }
 
-        private void OpenVisualizationTest(string fileName)
+        protected void OpenVisualizationTest(string fileName)
         {
             string relativePath = Path.Combine(
                 GetTestDirectory(ExecutingDirectory),
