+++ src/VisualizationTests/WpfVisualizationTests.csproj
@@ -113,6 +113,7 @@
       <Link>AssemblySharedInfo.cs</Link>
     </Compile>
     <Compile Include="CameraTests.cs" />
+    <Compile Include="CustomNodeTests.cs" />
     <Compile Include="HelixRenderPackageTests.cs" />
     <Compile Include="HelixWatch3DViewModelTests.cs" />
     <Compile Include="HelixWatch3DViewModelUnitTests.cs" />
